+++ packages/ragbits-document-search/src/ragbits/document_search/retrieval/rerankers/llm.py
@@ -5,10 +5,11 @@ from itertools import chain
 from pydantic import BaseModel
 from typing_extensions import Self
 
-from ragbits.core.audit import traceable
+from ragbits.core.audit.traces import traceable
 from ragbits.core.llms.base import LLM
 from ragbits.core.llms.litellm import LiteLLM, LiteLLMOptions
 from ragbits.core.prompt.prompt import Prompt
+from ragbits.core.types import NOT_GIVEN, NotGiven
 from ragbits.core.utils.config_handling import ObjectConstructionConfig, import_by_path
 from ragbits.document_search.documents.element import Element
 from ragbits.document_search.retrieval.rerankers.base import Reranker, RerankerOptions
@@ -39,20 +40,33 @@ class RerankerPrompt(Prompt[RerankerInput, str]):
     """
 
 
-class LLMReranker(Reranker[RerankerOptions]):
+class LLMRerankerOptions(RerankerOptions):
+    """
+    Object representing the options for the llm reranker.
+
+    Attributes:
+        top_n: The number of entries to return.
+        score_threshold: The minimum relevance score for an entry to be returned.
+        override_score: If True reranking will override element score.
+        llm_options: The options for the LLM.
+    """
+
+    llm_options: LiteLLMOptions | None | NotGiven = NOT_GIVEN
+
+
+class LLMReranker(Reranker[LLMRerankerOptions]):
     """
     Reranker based on LLM.
     """
 
-    options_cls = RerankerOptions
+    options_cls: type[LLMRerankerOptions] = LLMRerankerOptions
 
     def __init__(
         self,
         llm: LiteLLM,
         *,
         prompt: type[Prompt[RerankerInput, str]] | None = None,
-        llm_options: LiteLLMOptions | None = None,
-        default_options: RerankerOptions | None = None,
+        default_options: LLMRerankerOptions | None = None,
     ) -> None:
         """
         Initialize the LLMReranker instance.
@@ -60,7 +74,6 @@ class LLMReranker(Reranker[RerankerOptions]):
         Args:
             llm: The LLM instance to handle reranking.
             prompt: The prompt to use for reranking elements.
-            llm_options: The LLM options to override.
             default_options: The default options for reranking.
         """
         super().__init__(default_options=default_options)
@@ -71,12 +84,10 @@ class LLMReranker(Reranker[RerankerOptions]):
             logprobs=True,
             max_tokens=1,
             logit_bias={
-                self._llm.get_token_id("Yes"): 1,
-                self._llm.get_token_id("No"): 1,
+                self._llm.get_token_id(" Yes"): 1,
+                self._llm.get_token_id(" No"): 1,
             },
         )
-        if llm_options:
-            self._llm_options |= llm_options
 
     @classmethod
     def from_config(cls, config: dict) -> Self:
@@ -102,7 +113,7 @@ class LLMReranker(Reranker[RerankerOptions]):
         self,
         elements: Sequence[Sequence[Element]],
         query: str,
-        options: RerankerOptions | None = None,
+        options: LLMRerankerOptions | None = None,
     ) -> Sequence[Element]:
         """
         Rerank elements with LLM.
@@ -110,45 +121,52 @@ class LLMReranker(Reranker[RerankerOptions]):
         Args:
             elements: The elements to rerank.
             query: The query to rerank the elements against.
-            options: The RerankerOptions to use for reranking.
+            options: The options for reranking.
 
         Returns:
             The reranked elements.
         """
         merged_options = (self.default_options | options) if options else self.default_options
+        llm_options = (
+            self._llm_options | merged_options.llm_options if merged_options.llm_options else self._llm_options
+        )
 
         flat_elements = list(chain.from_iterable(elements))
-        scores = await self._score_elements(flat_elements, query)
+        scores = await self._score_elements(flat_elements, query, llm_options)
 
         scored_elements = list(zip(flat_elements, scores, strict=True))
         scored_elements.sort(key=lambda x: x[1], reverse=True)
 
         results = []
-        for element, score in scored_elements[: merged_options.top_n]:
+        for element, score in scored_elements[: merged_options.top_n or None]:
             if not merged_options.score_threshold or score >= merged_options.score_threshold:
                 if merged_options.override_score:
                     element.score = score
                 results.append(element)
         return results
 
-    async def _score_elements(self, elements: Sequence[Element], query: str) -> Sequence[float]:
+    async def _score_elements(
+        self,
+        elements: Sequence[Element],
+        query: str,
+        llm_options: LiteLLMOptions,
+    ) -> Sequence[float]:
         """
         Score the elements according to their relevance to the query using LLM.
 
         Args:
             elements: The elements to rerank.
             query: The query to rerank the elements against.
+            llm_options: The LLM options to use for scoring.
 
         Returns:
             The elements scores.
         """
-        merged_llm_options = self._llm.default_options | self._llm_options
-
         scores = []
         for element in elements:
             if element.text_representation:
                 prompt = self._prompt(RerankerInput(query=query, document=element.text_representation))
-                response = await self._llm.generate_with_metadata(prompt=prompt, options=merged_llm_options)
+                response = await self._llm.generate_with_metadata(prompt=prompt, options=llm_options)
                 prob = math.exp(response.metadata["logprobs"][0]["logprob"])
                 score = prob if response.content == "Yes" else 1 - prob
             else: