+++ packages/ragbits-evaluate/tests/unit/test_evaluator.py
@@ -3,6 +3,7 @@ from dataclasses import dataclass
 from typing import cast
 from unittest.mock import Mock
 
+import pytest
 from typing_extensions import Self
 
 from ragbits.core.utils.config_handling import ObjectConstructionConfig, WithConstructionConfig
@@ -30,17 +31,25 @@ class MockEvaluationTarget(WithConstructionConfig):
 
 
 class MockEvaluationPipeline(EvaluationPipeline[MockEvaluationTarget, MockEvaluationData, MockEvaluationResult]):
-    async def __call__(self, data: MockEvaluationData) -> MockEvaluationResult:
-        return MockEvaluationResult(
-            input_data=data.input_data,
-            processed_output=f"{self.evaluation_target.model_name}_{data.input_data}",
-            is_correct=data.input_data % 2 == 0,
-        )
+    async def __call__(self, data: Iterable[MockEvaluationData]) -> Iterable[MockEvaluationResult]:
+        return [
+            MockEvaluationResult(
+                input_data=row.input_data,
+                processed_output=f"{self.evaluation_target.model_name}_{row.input_data}",
+                is_correct=row.input_data % 2 == 0,
+            )
+            for row in data
+        ]
 
     @classmethod
     def from_config(cls, config: dict) -> "MockEvaluationPipeline":
         evaluation_target = WithConstructionConfig.subclass_from_config(config["evaluation_target"])
         return cls(evaluation_target=cast(MockEvaluationTarget, evaluation_target))
+
+
+class MockFailingEvaluationPipeline(EvaluationPipeline[MockEvaluationTarget, MockEvaluationData, MockEvaluationResult]):
+    async def __call__(self, data: Iterable[MockEvaluationData]) -> Iterable[MockEvaluationResult]:
+        raise Exception("This is a test exception")
 
 
 class MockDataLoader(DataLoader[MockEvaluationData]):
@@ -57,14 +66,23 @@ class MockDataLoader(DataLoader[MockEvaluationData]):
 
 
 class MockMetric(Metric[MockEvaluationResult]):
-    def compute(self, results: list[MockEvaluationResult]) -> dict:  # noqa: PLR6301
-        accuracy = sum(1 for r in results if r.is_correct) / len(results)
+    async def compute(self, results: list[MockEvaluationResult]) -> dict:  # noqa: PLR6301
+        accuracy = sum(1 for r in results if r.is_correct) / len(results) if results else 0
         return {"accuracy": accuracy}
 
 
-async def test_run_evaluation() -> None:
+@pytest.mark.parametrize(
+    ("pipeline_type", "expected_results", "expected_errors", "expected_accuracy"),
+    [(MockEvaluationPipeline, 4, 0, 0.5), (MockFailingEvaluationPipeline, 0, 1, 0)],
+)
+async def test_run_evaluation(
+    pipeline_type: type[EvaluationPipeline[MockEvaluationTarget, MockEvaluationData, MockEvaluationResult]],
+    expected_results: int,
+    expected_errors: int,
+    expected_accuracy: float,
+) -> None:
     target = MockEvaluationTarget(model_name="test_model")
-    pipeline = MockEvaluationPipeline(target)
+    pipeline = pipeline_type(target)
     dataloader = MockDataLoader()
     metrics = MetricSet(*[MockMetric()])
     evaluator = Evaluator()
@@ -72,19 +90,20 @@ async def test_run_evaluation() -> None:
     results = await evaluator.compute(
         pipeline=pipeline,
         dataloader=dataloader,
-        metrics=metrics,
+        metricset=metrics,
     )
 
-    assert len(results["results"]) == 4
-    assert 0 <= results["metrics"]["accuracy"] <= 1
-    assert all("test_model_" in r["processed_output"] for r in results["results"])
+    assert len(results.results) == expected_results
+    assert len(results.errors) == expected_errors
+    assert results.metrics["accuracy"] == expected_accuracy
+    assert all("test_model_" in r.processed_output for r in results.results)
 
 
 async def test_run_from_config() -> None:
     config = {
         "evaluation": {
             "dataloader": ObjectConstructionConfig.model_validate(
-                {"type": f"{__name__}:MockDataLoader", "config": {"dataset_size": 3}}
+                {"type": f"{__name__}:MockDataLoader", "config": {"dataset_size": 6}}
             ),
             "pipeline": {
                 "type": f"{__name__}:MockEvaluationPipeline",
@@ -101,5 +120,6 @@ async def test_run_from_config() -> None:
     }
     results = await Evaluator.run_from_config(config)
 
-    assert len(results["results"]) == 3
-    assert all("config_model_" in r["processed_output"] for r in results["results"])
+    assert len(results.results) == 6
+    assert results.metrics["accuracy"] == 0.5
+    assert all("config_model_" in r.processed_output for r in results.results)