+++ packages/ragbits-chat/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- updated ui build (#553)
+- api integration improvements + history context changes (#552)
+
 ## 0.17.1 (2025-05-09)
 
 ### Changed