+++ packages/ragbits-document-search/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+- Fix reranker options typing (#562)
+- Add query rephraser options (#560)
+- Rename DocumentMeta create_text_document_from_literal to from_literal (#561)
+- Update audit imports (#427)
+- BREAKING CHANGE: Adjust document search configurable interface (#554)
+- BREAKING CHANGE: Rename SearchConfig to DocumentSearchOptions (#554)
+- BREAKING CHANGE: Improve typing for SearchConfig (#554)
+
 ## 0.17.1 (2025-05-09)
 
 ### Changed